V1 Remaster
Images, videos?, plus souple

+++ create-ball3.sh
@@ -1,48 +1,36 @@
 #!/bin/bash
-#Pour inclure le décodeur dans les archives my-ball3.sh créées par create-ball3.sh 
-# il faut s'assurer que le fichier decodeur est également inclus dans l'archive. 
-#
-# Erreurs sur les images/videos
-[ $# -lt 0 ]&& exit 1 
-#Nous permet d'encoder notre fichier avant de l'écrire dans notre archiveur
-if test ! -e Encode.c
-then
-cat << 'encodeTMP' >> Encode.c
+# Script to create self-extracting archives with custom encoding
+
+[ $# -lt 1 ] && exit 1
+
+# Create encoder if needed
+if test ! -e Encode.c; then
+cat << 'encodeTMP' > Encode.c
 #include <stdio.h>
 #include <unistd.h>
 
-char CaraImprimable(unsigned char val) {
-    char Cara[32] = {'0', '1', '2', '3', '4', '5', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z'};
-    return Cara[val];
-}
+const char table[64] = {
+    'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P',
+    'Q','R','S','T','U','V','W','X','Y','Z','a','b','c','d','e','f',
+    'g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v',
+    'w','x','y','z','0','1','2','3','4','5','6','7','8','9','+','/'
+};
 
 void encode() {
-    char buffer[5];
-    int Lecture, flag;
+    unsigned char buffer[3];
+    int bytesRead;
 
-    fseek(stdin, 0, SEEK_END);
-    long tailleF = ftell(stdin);
-    fseek(stdin, 0, SEEK_SET);
+    while ((bytesRead = fread(buffer, 1, 3, stdin)) > 0) {
+        unsigned char char1 = buffer[0] >> 2;
+        unsigned char char2 = ((buffer[0] & 0x03) << 4) | (buffer[1] >> 4);
+        unsigned char char3 = ((buffer[1] & 0x0F) << 2) | (buffer[2] >> 6);
+        unsigned char char4 = buffer[2] & 0x3F;
 
-    if ((tailleF % 5) != 0) {
-        flag = 1;
+        putchar(table[char1]);
+        putchar(table[char2]);
+        putchar(bytesRead >= 2 ? table[char3] : '=');
+        putchar(bytesRead >= 3 ? table[char4] : '=');
     }
-
-    while ((Lecture = fread(buffer, 1, 5, stdin)) > 0) {
-        for (int k = 0; k < 8; ++k) {
-            char octetencode = 0;
-            for (int j = 0; j < Lecture && j < 5; ++j) {
-                octetencode = (octetencode << 1) | ((buffer[j] >> (7 - k)) & 1);
-            }
-
-            fprintf(stdout, "%c", CaraImprimable(octetencode));
-        }
-    }
-
-    if (flag == 1)
-        fprintf(stdout, "%c", '1');
-    for (int i = 0; i < 7; i++)
-        fprintf(stdout, "%c", '0');
 }
 
 int main() {
@@ -50,67 +38,42 @@ int main() {
     return 0;
 }
 encodeTMP
-gcc Encode.c -o encode
+gcc Encode.c -o encode || exit 1
 fi
 
-
-
-for j in $* 
-do 
-if [ ! -e my-ball3.sh ]
-then
-# on place le décodeur dans my-ball pour l'utiliser quand on en aura besoin
+# Initialize archive if needed
+if [ ! -e my-ball3.sh ]; then
 cat << 'decodeTMP' > my-ball3.sh
 #!/bin/bash
-cat << 'Decodage' >> Decode.c
+cat << 'Decodage' > Decode.c
 #include <stdio.h>
 
-char decodeChar(char c) {
-    // Table de correspondance des caractères imprimables
-    char Cara[32] = {'0', '1', '2', '3', '4', '5', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z'};
-
-    if (c == '\n') {
-        return '\n';
+int decodeChar(char c) {
+    const char table[64] = {
+        'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P',
+        'Q','R','S','T','U','V','W','X','Y','Z','a','b','c','d','e','f',
+        'g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v',
+        'w','x','y','z','0','1','2','3','4','5','6','7','8','9','+','/'
+    };
+    for (int i = 0; i < 64; ++i) {
+        if (c == table[i]) return i;
     }
-    for (int i = 0; i < 32; ++i) {
-        if (c == Cara[i]) {
-            return i;
-        }
-    }
-    // Caractère non trouvé
-    return 0;
+    return -1;
 }
 
 void decode() {
-    char buffer[8];
-    int fin = 0;
+    char buffer[4];
+    while (fread(buffer, 1, 4, stdin) == 4) {
+        int val1 = decodeChar(buffer[0]);
+        int val2 = decodeChar(buffer[1]);
+        int val3 = decodeChar(buffer[2]);
+        int val4 = decodeChar(buffer[3]);
 
-    while (fread(buffer, 1, 8, stdin) > 0) {
-        char octetsDecodes[5] = {0};
+        if (val1 == -1 || val2 == -1) continue;
 
-        if (buffer[0] == '1' && buffer[1] == '0' && buffer[2] == '0' && buffer[3] == '0' && buffer[4] == '0' && buffer[5] == '0' && buffer[6] == '0' && buffer[7] == '0') {
-            fin = 1;
-            break;
-        }
-
-        for (int k = 0; k < 5; ++k) {
-            char octetdecode = 0;
-            for (int j = 0; j < 8; ++j) {
-                octetdecode = (octetdecode << 1) | ((decodeChar(buffer[j]) >> k) & 1);
-            }
-
-            octetsDecodes[4 - k] = octetdecode;
-        }
-
-        if (fin == 1) {
-            break;
-        }
-
-        for (int i = 0; i < 5; ++i) {
-            if (octetsDecodes[i] != 0) {
-                fputc(octetsDecodes[i], stdout);
-            }
-        }
+        putchar((val1 << 2) | (val2 >> 4));
+        if (buffer[2] != '=') putchar((val2 << 4) | (val3 >> 2));
+        if (buffer[3] != '=') putchar((val3 << 6) | val4);
     }
 }
 
@@ -119,78 +82,40 @@ int main() {
     return 0;
 }
 Decodage
-gcc Decode.c -o decode
-chmod u+x decode
+gcc Decode.c -o decode && chmod +x decode || exit 1
 decodeTMP
 fi
 
-
-
-
-if [ -f $j ]
-then
-user=`ls -l $j | cut -d ' ' -f 1 | sed -E 's/.//' | sed 's/.../& /g'| cut -d ' ' -f 1`
-group=`ls -l $j | cut -d ' ' -f 1 | sed -E 's/.//' | sed 's/.../& /g'| cut -d ' ' -f 2`
-other=`ls -l $j | cut -d ' ' -f 1 | sed -E 's/.//' | sed 's/.../& /g'| cut -d ' ' -f 3`
-uno=$(expr `echo  $user | sed -E 's/./& + /g' | tr 'r' '4' | tr 'w' '2' | tr 'x' '1' | tr '-' '0'`0)
-dos=$(expr `echo  $group | sed -E 's/./& + /g' | tr 'r' '4' | tr 'w' '2' | tr 'x' '1' | tr '-' '0'`0)
-tres=$(expr `echo  $other | sed -E 's/./& + /g' | tr 'r' '4' | tr 'w' '2' | tr 'x' '1' | tr '-' '0'`0)
-#chmod contient la valeur octal des droit du fichier
-chmodfin=`echo $uno$dos$tres`
-#Nous archvions le fichier en l'encodant 
-#Nous utilisons mkdir -p pour creer son fichier parent sa saisie soit correct grace au mkdir -p
-cat <<TAG >> my-ball3.sh
-./decode > "$j" << TAG_D
-$(./encode -m < "$j")
+for j in "$@"; do
+    if [ -f "$j" ]; then
+        # permissions
+        perms=$(stat -c "%a" "$j")
+        
+        # Encode 
+        cat <<TAG >> my-ball3.sh
+./decode > "$j" << 'TAG_D'
+$(./encode < "$j")
 TAG_D
-chmod $chmodfin "$j"
+chmod $perms "$j"
 TAG
 
-
-fi
-
-#On vérifie d'abord si l'argument en entrée est un dossier avant de réaliser le parcours récursive
-if [ -d $j ]
-then
-#Tous les droits des dossiers (de forme octal) sont stocker dans les variable "uno" "dos" "tres" depuis le ls -ld 
-user=`ls -ld $j | cut -d ' ' -f 1 | sed -E 's/.//' | sed 's/.../& /g'| cut -d ' ' -f 1`
-group=`ls -ld $j | cut -d ' ' -f 1 | sed -E 's/.//' | sed 's/.../& /g'| cut -d ' ' -f 2`
-other=`ls -ld $j | cut -d ' ' -f 1 | sed -E 's/.//' | sed 's/.../& /g'| cut -d ' ' -f 3`
-uno=$(expr `echo  $user | sed -E 's/./& + /g' | tr 'r' '4' | tr 'w' '2' | tr 'x' '1' | tr '-' '0'`0)
-dos=$(expr `echo  $group | sed -E 's/./& + /g' | tr 'r' '4' | tr 'w' '2' | tr 'x' '1' | tr '-' '0'`0)
-tres=$(expr `echo  $other | sed -E 's/./& + /g' | tr 'r' '4' | tr 'w' '2' | tr 'x' '1' | tr '-' '0'`0)
-#chmod contient la valeur octal des droit du dossier
-chmoddossier=`echo $uno$dos$tres`
-
-
-
-cat << TAG >> my-ball3.sh 
-mkdir -p $j
-chmod $chmoddossier "$j"
+    elif [ -d "$j" ]; then
+        perms=$(stat -c "%a" "$j")
+        cat <<TAG >> my-ball3.sh
+mkdir -p "$j"
+chmod $perms "$j"
 TAG
-for i in $j/*
-do
-#Si le dossier contient un fichier, il est renvoyer dans la récusive   
-  	  $0 $i
- done
- fi
+        
+
+        find "$j" -exec "$0" {} \;
+    fi
 done
 
+chmod +x my-ball3.sh
+[ -e encode ] && rm encode Encode.c
 
-chmod u+x my-ball3.sh
-if test -e encode
-then
-rm encode
-rm Encode.c
-fi
-
-#Nous permetera de retirer le décodeur de notre fichier après l'exec 
-if test ! -e $1
-then
-cat << TAG >> my-ball3.sh
-rm decode
-rm Decode.c
+[ ! -e "$1" ] && cat <<TAG >> my-ball3.sh
+rm decode Decode.c
 TAG
-fi
-exit 0
 
+exit 0
